correct methods fib4() and fib5()

+++ src/main/java/Function.java
@@ -1,4 +1,5 @@
 import java.lang.reflect.Array;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Arrays;
 
@@ -6,11 +7,11 @@ import static java.lang.Math.*;
 
 public class Function {
     //version 4 - with array
-    public static BigInteger fib4(int number) {
+    public static BigDecimal fib4(int number) {
         if (number <= 1) {
-            return new BigInteger(String.valueOf(number));
+            return new BigDecimal(String.valueOf(number));
         } else {
-            BigInteger[] array = {new BigInteger("0"), new BigInteger("1")};
+            BigDecimal[] array = {new BigDecimal("0"), new BigDecimal("1")};
             for (int i = 2; i < number; i++) {
                 array[1] = array[1].add(array[0]);
                 array[0] = array[1].subtract(array[0]);
@@ -19,8 +20,8 @@ public class Function {
         }
     }
 
-    //version 5 - using the Binet formula
-    public static BigInteger fib5(int number) {
-        return new BigInteger(String.valueOf(round(pow((1 + sqrt(5))/2, number)/sqrt(5))));
+    //version 5 - using the Binet formula (на больших величинах теряется точность)
+    public static BigDecimal fib5(int number) {
+        return new BigDecimal(String.valueOf(round(pow((1 + sqrt(5))/2, number)/sqrt(5))));
     }
 }
